Rename Controller classes
